fix(advanced-color-picker): refactor color picker cell button
Improves accessibility by adding a `Button` which is clearly intended to open the
component. The current mechanism to open the component is with the use of a color picker
cell which acts as a button, this behaviour has been removed. The cell still remains
but is now placed next to the newly added `Button`, just to display the currently
chosen color to sighted users. An aria-describedby attribute has been added to the component
to establish a relationship between the newly added `Button` and a newly added description list
which details if a colour has been selected, and if it has, it describes the colours name for users
utilising assistive technology.

BREAKING CHANGE: The color picker cell which opens the components button behaviour
has been completely removed and replaced by a `Button` which is now rendered
within the component. As it was determined this method of opening the component
was not clear to users. Moreover,there was no mechanism to display the current
colour of the button to users using assistive technology. The data-element of
this cell remains the same, but its functionality now is completely superseded.